backend/app/utils/db.py
# async MongoDB connection using motor
from motor.motor_asyncio import AsyncIOMotorClient
import os
from dotenv import load_dotenv

load_dotenv()  # reads .env

MONGODB_URI = os.getenv("MONGODB_URI", "mongodb://localhost:27017")
DB_NAME = os.getenv("DB_NAME", "dataleak")

client: AsyncIOMotorClient | None = None
db = None

def connect_db():
    global client, db
    client = AsyncIOMotorClient(MONGODB_URI)
    db = client[DB_NAME]
    return db

def close_db():
    global client
    if client:
        client.close()
