backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
praw
tweepy
requests